framework: handle computed sem_index not being swap_chain_index
Right now we have a list of semaphores for the acquired images, as big
as the swapchain. Those semaphores are updated when we call
vkAcquireNextImageKHR. Right now we compute in advance the index that
this call will return to store it.

But this assumes that vkAcquireNextImageKHR will iterate through the
full size of the swapchain. It seems that on Wayland it returns the
first non-busy, and it is not using the full size.

This commit fix that by using the previously computed index as a
temporary place, and swapping later for the proper swap_chain_index
position.

FIXME: not really happy with this solution. Perhaps tracking with a
different index for the semaphore.